Remove date from article URLs
This is because I want the liberty to change the date on posts
if I choose to, e.g. because I update them after publishing.
Also, the URLs are just nicer. You don't need the date in the URL.

To do this, I had to stop using the native "posts" collection
and make my own collection called "articles". Then move all
markdown files there, and create "date" fields for sorting,
and use the "redirect-from" plugin to redirect the old URLs
to the new.